+++ apps/guardian/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Make it possible to override the node endpoint to use through `--node` command line argument
+
 ## 0.1.2
 
 - Updates concordium rust dependencies